fix: normalize bundle when generating CCD
Ref. metriport/metriport-internal#799
